Moved SparkTestUtils to package for test utils
